New graft server main test added; first step.
